Edit lexer.c

+++ src/libgeometry/lexer.c
@@ -1,9 +1,12 @@
+#include <stdlib.h>
+#include <string.h>
+
 extern int SIZE; 
 int checker(char* str)
 {
     int ret = 1;
     char* rec = (char*)malloc(SIZE * sizeof(char));
-    for (int i = 0; i < strlen(str); i++) {
+    for (int i = 0; i < (int) strlen(str); i++) {
         if (str[i] != '(')
             rec[i] = str[i];
         else
